v3.1 - See CHANGELOG.md

+++ setup.py
@@ -16,7 +16,15 @@ target_directory = (
 # Copy the config.yml file to the target directory if it exists
 if target_directory and os.path.isfile("config.yml"):
     os.makedirs(target_directory, exist_ok=True)
-    shutil.copy("config.yml", target_directory)
+    # If file already exists, create a new one
+    if os.path.isfile(target_directory+'/config.yml'):
+        print('The file '+target_directory+'/config.yml already exists.\nCreating config.yml.NEW but leaving existing config.\nIf you need the new file, then remove the current one and rename config.yml.NEW to config.yml\n')
+        os.rename(target_directory+'/config.yml',target_directory+'/config.yml.OLD')
+        shutil.copy("config.yml", target_directory)
+        os.rename(target_directory+'/config.yml',target_directory+'/config.yml.NEW')
+        os.rename(target_directory+'/config.yml.OLD',target_directory+'/config.yml')
+    else:
+        shutil.copy("config.yml", target_directory)
     
 setup(
     name="knoxnl",
